requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
cycler==0.11.0
dask==2022.2.0
decorator==5.1.1
draggan==1.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
ffmpy==0.3.1
filelock==3.12.2
fire==0.5.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.5.4
google-pasta==0.2.0
gradio==3.34.0
gradio_client==0.2.6
grpcio==1.56.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.4
linkify-it-py==2.0.2
locket==1.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
networkx==2.6.3
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==1.3.5
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
protobuf==3.20.0
ptyprocess==0.7.0
pydantic==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.26.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-apt==1.6.6
python-dateutil==2.8.2
python-distutils-extra==2.39
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.31.0
scikit-image==0.15.0
scikit-learn==1.0.2
scipy==1.1.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
unattended-upgrades==0.1
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
